REF: example/: renamed `example_plot.py` to `_example_plot.py`.
