+++ examples/1_basic.py
@@ -3,7 +3,7 @@ Example 1: Analysing a single file with hydrodynamic output data.
 
 Authors: Soesja Brunink & Gijs G. Hendrickx
 """
-from examples.example_plot import create_figure
+from examples._example_plot import create_figure
 from src.processing import map_ecotopes
 
 # map ecotopes based on hydrodynamic output data: 'file-name_map.nc'
