+++ examples/5_config.py
@@ -10,7 +10,7 @@ expected in the same directory (i.e., `wd_config`).
 
 Authors: Soesja Brunink & Gijs G. Hendrickx
 """
-from examples.example_plot import create_figure
+from examples._example_plot import create_figure
 from src.processing import map_ecotopes
 
 # map ecotopes based on hydrodynamic output data: 'file-name_map.nc'
