+++ examples/_example_plot.py
